refactor(api-client-auth): rename interceptor to authenticate
