+++ packages/api-client-auth/src/interceptors/hmac-auth-interceptor/index.ts
